move grouped_index_vector.hpp to common
