+++ power_grid_model_c/power_grid_model/include/power_grid_model/common/grouped_index_vector.hpp
